docconnect-backend-main/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
boto3==1.18.31
botocore==1.21.31
certifi==2023.5.7
charset-normalizer==3.1.0
Django==3.2.6
django-cors-headers==3.10.0
django-extensions==3.2.1
django-filter==2.4.0
django-ses==2.3.1
django-storages==1.11.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
future==0.18.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.0.0
jmespath==0.10.0
Markdown==3.4.1
numpy==1.21.4
pandas==1.3.4
Pillow==8.3.1
psycopg2-binary==2.9.2
PyJWT==1.7.1
pyotp==2.6.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.31.0
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
tzdata==2022.6
urllib3==1.26.6
zipp==3.11.0
